+++ app/Http/Controllers/CategoryController.php
@@ -167,22 +167,17 @@ class CategoryController extends Controller
         $categoria = Category::findOrFail($request->id);//Se obtiene el objeto producto por el id
 
         if($request->hasFile('image')){//Comprobar si existe la imagen y no tenga valor null
-            if(isset($categoria->category_thumbnail) && !is_null($categoria->category_thumbnail)){
-                $imagenEliminar = str_replace('storage', 'public', $categoria->category_thumbnail);//reemplazar la palabra storage por public
-                Storage::delete($imagenEliminar); //Eliminar la imagen actual de la categoria
-            }
-
-            $extensionImagen = '.'.$request->file('image')->extension();
-            $nombreSinExtension = trim($request->category_thumbnail, $extensionImagen);
-            $nombreFinal = $nombreSinExtension.$extensionImagen;
-            $imagen = $request->file('image')->storeAs('public/imagenes', $nombreFinal);//Obtener la ruta temporal de la imagen y cambiar el nombre y almacenar en 'public/imagenes'
-            $url = Storage::url($imagen);//Guardar la imagen en el Storage
+            $url = $this->agregarImagen($request, $categoria);
         }else{
-            if(isset($categoria->category_thumbnail) && !is_null($categoria->category_thumbnail)){
-                $imagenEliminar = str_replace('storage', 'public', $categoria->category_thumbnail);//reemplazar la palabra storage por public
-                Storage::delete($imagenEliminar); //Eliminar la imagen actual de la categoria
-            }
             $url="";
+            if(isset($request->category_thumbnail) && !is_null($request->category_thumbnail)){
+                $url = $request->category_thumbnail;
+            }else{
+                if(isset($categoria->category_thumbnail) && !is_null($categoria->category_thumbnail)){
+                    $imagenEliminar = str_replace('storage', 'public', $categoria->category_thumbnail);//reemplazar la palabra storage por public
+                    Storage::delete($imagenEliminar); //Eliminar la imagen actual de la categoria
+                }
+            }
         }
 
         $categoria->category_name = $request->category_name;
@@ -212,6 +207,21 @@ class CategoryController extends Controller
                 'status' => $_ENV['CODE_STATUS_SERVER_ERROR']
             ]);
         }
+    }
+
+    function agregarImagen(Request $request, Category $categoria)
+    {
+        if(isset($categoria->category_thumbnail) && !is_null($categoria->category_thumbnail)){
+            $imagenEliminar = str_replace('storage', 'public', $categoria->category_thumbnail);//reemplazar la palabra storage por public
+            Storage::delete($imagenEliminar); //Eliminar la imagen actual de la categoria
+        }
+
+        $extensionImagen = '.'.$request->file('image')->extension();
+        $nombreSinExtension = trim($request->category_thumbnail, $extensionImagen);
+        $nombreFinal = $nombreSinExtension.$extensionImagen;
+        $imagen = $request->file('image')->storeAs('public/imagenes', $nombreFinal);//Obtener la ruta temporal de la imagen y cambiar el nombre y almacenar en 'public/imagenes'
+        $url = Storage::url($imagen);//Guardar la imagen en el Storage
+        return $url;
     }
 
     /**
